refactor: combine throttle test file
